move modals to walletconnect store
